refactor: eliminate separate sql command module
